# docs/swagger.yaml
basePath: /
definitions:
...
  data.TokenType:
    enum:
    - fungible
    - non-fungible
    type: string
    x-enum-varnames:
    - FUNGIBLE
    - NONFUNGIBLE
...
  github_com_quantum-bridge_core_cmd_service_shared.Token:
    properties:
      attributes:
        allOf:
        - $ref: '#/definitions/shared.TokenAttributes'
        description: Attributes is the attributes of the token entity.
      id:
        description: ID is the identifier of the entity.
        type: string
      relationships:
        allOf:
        - $ref: '#/definitions/shared.TokenRelationships'
        description: Relationships is the relationships of the token entity.
      type:
        allOf:
        - $ref: '#/definitions/shared.EntityType'
        description: Type is the type of the entity.
    required:
    - attributes
    - id
    - relationships
    - type
    type: object
...
  requests.WithdrawDTO:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/requests.WithdrawRequest'
        description: Data is the data of the withdrawal request.
    required:
    - data
    type: object
  requests.WithdrawRequest:
    properties:
      chain_from:
        description: ChainFrom is the source chain ID.
        type: string
      from:
        description: From is the address of the sender in the destination chain. Should
          be used only if the sender address is different with source chain.
        type: string
      token_id:
        description: TokenID is the token ID of the token.
        type: string
      tx_hash:
        description: TxHash is the hash of the transaction in the source chain that
          locked the token.
        type: string
    required:
    - chain_from
    - token_id
    - tx_hash
    type: object
...
  shared.EntityType:
    enum:
    - balance
    - chain
    - evm_transaction
    - nft
    - processed_transaction
    - token
    type: string
    x-enum-varnames:
    - BALANCE
    - CHAIN
    - EVM_TRANSACTION
    - NFT
    - PROCESSED_TRANSACTION
    - TOKEN
  shared.Key:
    properties:
      id:
        description: ID is the identifier of the entity.
        type: string
      type:
        allOf:
        - $ref: '#/definitions/shared.EntityType'
        description: Type is the type of the entity.
    required:
    - id
    - type
    type: object
...
  shared.RelationCollection:
    properties:
      data:
        description: Data is list of Key objects.
        items:
          $ref: '#/definitions/shared.Key'
        type: array
    required:
    - data
    type: object
  shared.TokenAttributes:
    properties:
      icon:
        description: Icon is the icon of the token.
        type: string
      name:
        description: Name is the name of the token.
        type: string
      symbol:
        description: Symbol is the symbol of the token.
        type: string
      token_type:
        allOf:
        - $ref: '#/definitions/data.TokenType'
        description: TokenType is the type of the token.
    required:
    - name
    - symbol
    - token_type
    type: object
  shared.TokenListResponse:
    properties:
      data:
        description: Data is the list of tokens.
        items:
          $ref: '#/definitions/github_com_quantum-bridge_core_cmd_service_shared.Token'
        type: array
      included:
        description: Included is the included object in the response.
        items:
          items:
            type: integer
          type: array
        type: array
    required:
    - data
    type: object
  shared.TokenRelationships:
    properties:
      chains:
        allOf:
        - $ref: '#/definitions/shared.RelationCollection'
        description: Chains is the chains that are used by the token.
    required:
    - chains
    type: object
  shared.TransactionsResponse:
    properties:
      data:
        description: Data is the body of transaction that is returned in the response.
      included:
        description: Included is the included chain data of the response.
        items:
          items:
            type: integer
          type: array
        type: array
    required:
    - data
    type: object
host: localhost:8000
info:
  contact: {}
  description: Core bridge API is a service that responsible for the communication
    between blockchains.
  title: Core bridge API
  version: "1.0"
paths:
...
  /tokens:
    get:
      consumes:
      - application/json
      description: Get the list of tokens based on the filter type and include chains
        flag.
      operationId: getTokens
      parameters:
      - description: 'Include chains in the response. Items Value: [`true`, `false`]'
        in: path
        name: include_chains
        type: boolean
      - description: 'Filter by chain type. Items Value: [`''chain''`]'
        in: query
        name: filter[token_type]
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/shared.TokenListResponse'
        "400":
          description: Bad request
        "500":
          description: Internal server error
      summary: Get Tokens
      tags:
      - Tokens
...
  /transfers/withdraw:
    post:
      consumes:
      - application/json
      description: Check if lock transaction is valid and withdraw the token from
        the bridge. Returns the transaction
      operationId: withdraw
      parameters:
      - description: Request body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/requests.WithdrawDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/shared.TransactionsResponse'
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Internal server error
      summary: Withdraw
      tags:
      - Transfers
schemes:
- http
swagger: "2.0"